requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
PyJWT==2.8.0
python-dotenv==1.0.0
